Updated and renamed object answer page
