+++ lib/ui/main_reports/view_report/fill_answer/view_location_page.dart
@@ -1,5 +1,6 @@
 import 'package:checkup_app/data/data_master.dart';
 import 'package:checkup_app/models/location.dart';
+import 'package:checkup_app/models/location_answer.dart';
 import 'package:checkup_app/models/report_answer.dart';
 import 'package:checkup_app/ui/main_reports/view_report/fill_answer/fill_location_page.dart';
 import 'package:flutter/material.dart';
@@ -20,6 +21,9 @@ class ViewLocationPage extends StatefulWidget {
 }
 
 class _ViewLocationPageState extends State<ViewLocationPage> {
+  LocationAnswer get locationAnswer =>
+      widget.reportAnswer.getOrCreateLocationAnswer(widget.location);
+
   @override
   Widget build(BuildContext context) {
     return Scaffold(
@@ -28,19 +32,22 @@ class _ViewLocationPageState extends State<ViewLocationPage> {
         children: [
           buildLocationHeader(),
           //TODO: Add new redesign widgets
+          buildIssueReporter(),
           Padding(
             padding: const EdgeInsets.all(16.0),
-            child: ElevatedButton(
-                onPressed: () {
-                  Navigator.push(
-                      context,
-                      MaterialPageRoute(
-                        builder: (context) => FillLocationPage(
-                            answer: widget.reportAnswer,
-                            location: widget.location),
-                      ));
-                },
-                child: const Text("View checks")),
+            child: ElevatedButton.icon(
+              onPressed: () {
+                Navigator.push(
+                    context,
+                    MaterialPageRoute(
+                      builder: (context) => FillLocationPage(
+                          answer: widget.reportAnswer,
+                          location: widget.location),
+                    ));
+              },
+              label: const Text("View checks"),
+              icon: Icon(Icons.exit_to_app),
+            ),
           )
         ],
       ),
@@ -49,7 +56,7 @@ class _ViewLocationPageState extends State<ViewLocationPage> {
 
   Widget buildLocationHeader() {
     return ListTile(
-      leading: const Icon(Icons.place, size: 72),
+      leading: const Icon(Icons.place, size: 64),
       title: Text(
         widget.location.name,
         textScaleFactor: 1.5,
@@ -59,9 +66,41 @@ class _ViewLocationPageState extends State<ViewLocationPage> {
           Map<String, int> info =
               widget.location.getInfo(widget.reportAnswer, dm);
           return Text(
-              "${info['checked']}/${info['total']} checked, ${info['issues']} issue(s)");
+              "${info['checked']}/${info['total']} checked\n${info['issues']} issue(s)");
         },
       ),
+    );
+  }
+
+  Widget buildIssueReporter() {
+    return Column(
+      children: [
+        SwitchListTile(
+          value: !locationAnswer.status,
+          title: Text("Report issue"),
+          subtitle: Text("Use this if you found location issues"),
+          onChanged: (value) {
+            setState(() {
+              locationAnswer.status = !value;
+              // Provider.of<DataMaster>(context, listen: false).update();
+            });
+          },
+        ),
+        if (!locationAnswer.status)
+          Padding(
+            padding: EdgeInsets.all(16.0),
+            child: Builder(builder: (context) {
+              return TextField(
+                controller: TextEditingController(text: locationAnswer.notes),
+                decoration: InputDecoration(labelText: 'Issue description'),
+                onChanged: (value) {
+                  locationAnswer.notes = value;
+                  //TODO: Why issues and notes? Why not a single thing?
+                },
+              );
+            }),
+          )
+      ],
     );
   }
 }
