+++ lib/ui/main_reports/view_report/fill_answer/widgets/custom_add_issue_tile.dart
@@ -1,0 +1,33 @@
+import 'package:flutter/material.dart';
+
+class CustomAddIssueTile extends StatefulWidget {
+  const CustomAddIssueTile({super.key, required this.onCreateIssue});
+  final Function(String name) onCreateIssue;
+
+  @override
+  State<CustomAddIssueTile> createState() => _CustomAddIssueTileState();
+}
+
+class _CustomAddIssueTileState extends State<CustomAddIssueTile> {
+  late TextEditingController textEditingController;
+  @override
+  void initState() {
+    textEditingController =
+        TextEditingController.fromValue(TextEditingValue.empty);
+    super.initState();
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return ListTile(
+      title: TextField(
+        controller: textEditingController,
+        decoration: const InputDecoration(hintText: 'Other issue'),
+        onSubmitted: (value) {
+          widget.onCreateIssue(value);
+          textEditingController.text = "";
+        },
+      ),
+    );
+  }
+}
